requirements.txt
fastapi
gunicorn
uvicorn
jinja2
pytest
requests
pre-commit
python-dotenv
pydantic-settings
aiofiles
python-multipart
pillow
pytesseract